Re-add the stat

+++ entrypoint.sh
@@ -98,10 +98,9 @@ rsync --rsh="ssh -v -p 22 -i ${WPE_SSHG_KEY_PRIVATE_PATH} -o StrictHostKeyChecki
 if [[ -n ${SCRIPT} || -n ${CACHE_CLEAR} ]]; then
 
     if [[ -n ${SCRIPT} ]]; then
-      # ssh -v -p 22 -i ${WPE_SSHG_KEY_PRIVATE_PATH} -o StrictHostKeyChecking=no -o ControlPath="$SSH_PATH/ctl/%C" $WPE_FULL_HOST "stat sites/${WPE_ENV_NAME}/${INPUT_SCRIPT}"
+      ssh -v -p 22 -i ${WPE_SSHG_KEY_PRIVATE_PATH} -o StrictHostKeyChecking=no -o ControlPath="$SSH_PATH/ctl/%C" $WPE_FULL_HOST "stat sites/${WPE_ENV_NAME}/${INPUT_SCRIPT}"
 
-      # if [[ $? -ne 0 && -f ${INPUT_SCRIPT} ]]; then
-      if [[ -f ${INPUT_SCRIPT} ]]; then
+      if [[ $? -ne 0 && -f ${INPUT_SCRIPT} ]]; then
         ssh -v -p 22 -i ${WPE_SSHG_KEY_PRIVATE_PATH} -o StrictHostKeyChecking=no -o ControlPath="$SSH_PATH/ctl/%C" $WPE_FULL_HOST "mkdir -p sites/${WPE_ENV_NAME}/$(dirname ${INPUT_SCRIPT})"
 
         rsync --rsh="ssh -v -p 22 -i ${WPE_SSHG_KEY_PRIVATE_PATH} -o StrictHostKeyChecking=no -o 'ControlPath=$SSH_PATH/ctl/%C'" ${INPUT_SCRIPT} wpe_gha+"$WPE_SSH_USER":sites/"$WPE_ENV_NAME"/$(dirname ${INPUT_SCRIPT})
